change: move test/stories dir
