+++ components/__stories__/ArticleHeadInfos.stories.ts
@@ -1,0 +1,13 @@
+import ArticleHeadInfos from '../ArticleHeadInfos';
+
+export default {
+  title: 'component/ArticleHeadInfos',
+  component: ArticleHeadInfos,
+};
+
+export const Default = {
+  args: {
+    published: '2022-01-01T00:00:00.000Z',
+    category: ['本', 'プログラム'],
+  }
+};
